+++ cli/azd/pkg/infra/provisioning/manager.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"fmt"
 	"io"
-	"log"
 	"strings"
 
 	"github.com/azure/azure-dev/cli/azd/pkg/azureutil"
@@ -179,10 +178,6 @@ func (m *Manager) deploy(
 			go func() {
 				for progress := range deployTask.Progress() {
 					m.updateSpinnerTitle(spinner, progress.Message)
-
-					if m.formatter.Kind() == output.JsonFormat {
-						m.writeJsonOutput(ctx, progress.Operations)
-					}
 				}
 			}()
 
@@ -201,10 +196,6 @@ func (m *Manager) deploy(
 
 	if err != nil {
 		return nil, fmt.Errorf("error deploying infrastructure: %w", err)
-	}
-
-	if m.formatter.Kind() == output.JsonFormat {
-		m.writeJsonOutput(ctx, deployResult.Operations)
 	}
 
 	m.console.Message(ctx, output.WithSuccessFormat("\nAzure resource provisioning completed successfully"))
@@ -271,6 +262,7 @@ func (m *Manager) ensureLocation(ctx context.Context, deployment *Deployment) (s
 		// project.
 		selected, err := azureutil.PromptLocation(
 			ctx,
+			m.env,
 			"Please select an Azure location to use to store deployment metadata:",
 		)
 		if err != nil {
@@ -291,7 +283,7 @@ func (m *Manager) runAction(
 ) error {
 	var spinner *spin.Spinner
 
-	if interactive {
+	if interactive && (m.formatter == nil || m.formatter.Kind() != output.JsonFormat) {
 		spinner, ctx = spin.GetOrCreateSpinner(ctx, m.console.Handles().Stdout, title)
 		defer spinner.Stop()
 		defer m.console.SetWriter(nil)
@@ -310,13 +302,6 @@ func (m *Manager) updateSpinnerTitle(spinner *spin.Spinner, message string) {
 	}
 
 	spinner.Title(fmt.Sprintf("%s...", message))
-}
-
-func (m *Manager) writeJsonOutput(ctx context.Context, output any) {
-	err := m.formatter.Format(output, m.writer, nil)
-	if err != nil {
-		log.Printf("error formatting output: %s", err.Error())
-	}
 }
 
 // Monitors the interactive channel and starts/stops the terminal spinner as needed